Add an @uri to the root-level res:resources element of resources.
Since the resource tree is supposed to contain all resources used for building
a website, it does make sense to consider that the root element will be the one
used to generate the top-level page of the website.  Configuring whatever we
want in the mapping.xml file was quite fine, but also quite buggy by-design.