Rework the main-navigation plugin.
Since we know that the top-level resource element is the top-level page of the
website, it's a piece of cake to find the top-level links.  While here, add the
ability to add a (redundant) link to the website home page.  Sometimes it's
useful (it mostly depends of the pages presentation).